Moved the lexical document builders into @tryghost/test-data
no ref

- the post builder's private paragraph-doc helper and e2e's lexical.ts
  (card-spec builder used by createWithCards) both constructed the same
  Lexical root-document scaffolding; the package now owns one envelope
  with the paragraph and card-spec forms built on it
- e2e consumes buildLexical/buildLexicalParagraph/CardSpec via the
  @/data-factory barrel, which also let the post-visibility spec drop
  its own copy of the paragraph scaffolding
